feat: add custom login and 404 pages
- Remove parent orgs route since we only have one org
- Remove orgs page as well
