+++ src/pages/404.tsx
@@ -1,8 +1,7 @@
 import { MarkGithubIcon } from '@primer/octicons-react'
-import { Box, Button, Octicon, Text } from '@primer/react'
-import { signIn } from 'next-auth/react'
+import { Box, Octicon, Text } from '@primer/react'
 
-function LoginBox() {
+function Custom404() {
   return (
     <Box
       sx={{
@@ -18,7 +17,6 @@ function LoginBox() {
         borderColor: 'border.default',
         borderRadius: '12px',
         padding: '40px',
-        marginTop: '60px',
       }}
     >
       <Box sx={{ marginBottom: '15px' }}>
@@ -31,29 +29,19 @@ function LoginBox() {
           alignItems: 'center',
         }}
       >
-        <Box sx={{ marginBottom: '15px' }}>
+        <Box>
           <Box sx={{ textAlign: 'center' }}>
-            <Text sx={{ fontSize: '3' }}>Sign in to get started</Text>
+            <Text sx={{ fontSize: '3' }}>404</Text>
           </Box>
           <Box sx={{ textAlign: 'center' }}>
             <Text sx={{ fontSize: '1', color: 'fg.muted' }}>
-              Internal Contribution Forks
+              Page not found
             </Text>
           </Box>
-        </Box>
-        <Box>
-          <Button
-            variant="primary"
-            onClick={() => {
-              signIn('github')
-            }}
-          >
-            Sign in with GitHub
-          </Button>
         </Box>
       </Box>
     </Box>
   )
 }
 
-export default LoginBox
+export default Custom404
